LPS-88156 FragmentsEditor utils separated in different files
