+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/fragment_entry_link/FragmentEntryLinkList.es.js
@@ -20,7 +20,8 @@ import {
 	DROP_TARGET_BORDERS,
 	DROP_TARGET_ITEM_TYPES
 } from '../../reducers/placeholders.es';
-import {getFragmentRowIndex, setIn} from '../../utils/utils.es';
+import {getFragmentRowIndex} from '../../utils/FragmentsEditorGetUtils.es';
+import {setIn} from '../../utils/FragmentsEditorUpdateUtils.es';
 import state from '../../store/state.es';
 import templates from './FragmentEntryLinkList.soy';
