+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/reducers/fragments.es.js
@@ -10,14 +10,16 @@ import {
 } from '../components/fragment_entry_link/FragmentEntryLinkContent.es';
 import {
 	add,
-	getColumn,
-	getDropSectionPosition,
-	getFragmentColumn,
 	remove,
 	setIn,
 	updateIn,
 	updateLayoutData
-} from '../utils/utils.es';
+} from '../utils/FragmentsEditorUpdateUtils.es';
+import {
+	getColumn,
+	getDropSectionPosition,
+	getFragmentColumn
+} from '../utils/FragmentsEditorGetUtils.es';
 
 /**
  * @param {!object} state
